requirements.txt
Flask
requests
python-slugify
